Rework authentication flow
I was able to identify that the issue with the admin panel pulling the same content has to do with a bearer token being undefined at the point of the call. The getAccessToken() function should've taken care of this condition, but it wasn't. I've refactored it to fit the flow oauth requires.